xonsh: activation moved from profile to bash_profile
the activation if xonsh in `~/.profile` creates a problem with login
from lightdm. im sure its because if the `exec` in the profile. it
executed xonsh instead of `i3` for example. moving to `~/.bash_profile`
doesnt work well with X. that file never gets executed. i source the
bash_profile through bashrc too. like this it works with terminals and
over ssh.